Output dimple mtz to main output directory
Moves around the code for checking pdb similarity to avoid making the
calculate maps step responsible for copying the dimple mtz file.
